Système de Recommandation Multimodal/tp_nlp/requirements.txt
clip-by-openai
ipywidgets
jupyter
matplotlib
notebook
numpy
pandas
pillow
scikit-learn
seaborn
sentence-transformers
streamlit
torch
tokenizers
torchvision
tqdm
transformers[torch]
umap-learn
